chore: remove legacy sect code from ui
