+++ src/features/index.js
@@ -2,11 +2,13 @@
 // Add feature UI mounts here as features migrate.
 
 import { mountProficiencyUI } from "./proficiency/ui/weaponProficiencyDisplay.js";
+import { mountSectUI } from "./sect/ui/sectScreen.js";
 // Example placeholder for later:
 // import { mountWeaponGenUI } from "./weaponGeneration/ui/weaponGenerationDisplay.js";
 
 export function mountAllFeatureUIs(state) {
   mountProficiencyUI(state);
+  mountSectUI(state);
   // mountWeaponGenUI?.(state);
 }
