+++ src/features/sect/state.js
@@ -1,0 +1,18 @@
+export const sectState = {
+  buildings: {},
+  bonuses: {
+    qiRegenMult: 0,
+    qiCapMult: 0,
+    herbYield: 0,
+    oreYield: 0,
+    woodYield: 0,
+    alchemySlots: 0,
+    alchemySuccess: 0,
+    atkBase: 0,
+    defBase: 0,
+    disciples: 0,
+    lawPoints: 0,
+    breakthroughBonus: 0,
+    foundationMult: 0,
+  },
+};
