+++ src/game/state.js
@@ -135,14 +135,6 @@ export const defaultState = () => {
       alchemy: {}
     }
   },
-  // Sect Buildings System
-  buildings: {}, // Building levels: {building_key: level}
-  // Building bonuses (calculated from building levels)
-  buildingBonuses: {
-    qiRegenMult: 0, qiCapMult: 0, herbYield: 0, oreYield: 0, woodYield: 0,
-    alchemySlots: 0, alchemySuccess: 0, atkBase: 0, defBase: 0,
-    disciples: 0, lawPoints: 0, breakthroughBonus: 0, foundationMult: 0
-  }
   };
 };
